crypto: kecacc: Use HashType instead of raw u8

+++ src/server/crypto/hashmux/src/hashmux.rs
@@ -665,7 +665,7 @@ impl HashMuxReceiver {
             KECACC.start_load(&state[to]);
         }
         else {
-            KECACC.start_init(sess.algo.unwrap().ty as u8);
+            KECACC.start_init(sess.algo.unwrap().ty);
         }
     }
 }
